Fixed botcmd.txt
- Fixed error with botcmd.txt being locked when server is running
- Moved test functions to dedicated DebugFunctions.cs
